move the create car request function to carRequest
